# api-documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Sosmet - Backend API Documentation",
    "contact": {}
  },
  "host": "localhost:8000",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/new-post-image/{user_id}": {
      "post": {
        "summary": "Create Post Image",
        "tags": ["Post"],
        "operationId": "CreatePostImage",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
...
    "/upload-image/{user_id}": {
      "post": {
        "summary": "Upload Image",
        "tags": ["Storage"],
        "operationId": "UploadImage",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
    "/upload-video/{user_id}": {
      "post": {
        "summary": "Upload Video",
        "tags": ["Storage"],
        "operationId": "UploadVideo",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "video",
            "in": "formData",
            "type": "file",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
...
    "/storage-videos/{user_id}": {
      "get": {
        "summary": "Fetch Videos By User Id",
        "tags": ["Storage"],
        "operationId": "FetchVideos",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
    "/storage-docs/{user_id}": {
      "get": {
        "summary": "Fetch Docs By User Id",
        "tags": ["Storage"],
        "operationId": "FetchDocs",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
...
  },
...
}